refactor: export types and status of models to constants
