+++ src/constants.js
@@ -6,7 +6,26 @@ WHITELISTEAD_POLL_ACCOUNT_NUMBERS = [
 
 MAX_GOVERNANCE_VOTE_PER_ACCOUNT = 3;
 
+const PollStatus = {
+  IN_PROGRESS: 0,
+  COMPLETED: 1,
+  CANCELLED: 2,
+};
+
+const UserType = {
+  GENERAL: "GENERAL",
+  GOVERNER: "GOVERNER",
+};
+
+const VoteType = {
+  GOVERNANCE: "GOVERNANCE",
+  POLL: "POLL",
+};
+
 module.exports = {
   WHITELISTEAD_POLL_ACCOUNT_NUMBERS,
   MAX_GOVERNANCE_VOTE_PER_ACCOUNT,
+  PollStatus,
+  UserType,
+  VoteType,
 };
